+++ crates/fluence-app-service/src/config.rs
@@ -23,6 +23,8 @@ use std::path::PathBuf;
 #[derive(Default)]
 pub struct AppServiceConfig<WB: WasmBackend> {
     /// Used for preparing filesystem on the service initialization stage.
+    pub service_working_dir: PathBuf,
+    /// Location for /tmp and /local dirs.
     pub service_base_dir: PathBuf,
     pub marine_config: MarineConfig<WB>,
 }