+++ crates/fluence-app-service/src/service.rs
@@ -24,6 +24,7 @@ use marine_wasm_backend_traits::WasmBackend;
 #[cfg(feature = "raw-module-api")]
 use marine_wasm_backend_traits::WasiState;
 use marine::Marine;
+use marine::MarineModuleConfig;
 use marine::IValue;
 
 use serde_json::Value as JValue;
@@ -31,7 +32,7 @@ use maplit::hashmap;
 
 use std::convert::TryInto;
 use std::collections::HashMap;
-use std::path::PathBuf;
+use std::path::Path;
 use std::io::ErrorKind;
 
 const SERVICE_ID_ENV_NAME: &str = "service_id";
@@ -126,8 +127,9 @@ impl<WB: WasmBackend> AppService<WB> {
 
     /// Prepare service before starting by:
     ///  1. creating a directory structure in the following form:
-    ///     - service_base_dir/service_id/SERVICE_LOCAL_DIR_NAME
-    ///     - service_base_dir/service_id/SERVICE_TMP_DIR_NAME
+    ///     - service_tmp_dir/service_id/SERVICE_LOCAL_DIR_NAME
+    ///     - service_tmp_dir/service_id/SERVICE_TMP_DIR_NAME
+    ///  2. rooting all mapped and preopened directories at service_working_dir
     ///  2. adding service_id to environment variables
     ///  3. moving all the user defined mapped dirs and preopened files to service_base_dir/service_id/
     fn set_env_and_dirs(
@@ -135,25 +137,24 @@ impl<WB: WasmBackend> AppService<WB> {
         service_id: String,
         mut envs: HashMap<Vec<u8>, Vec<u8>>,
     ) -> Result<()> {
-        let create = |dir: &PathBuf| match std::fs::create_dir(dir) {
-            Err(e) if e.kind() == ErrorKind::AlreadyExists => Ok(()),
-            Err(err) => Err(AppServiceError::CreateDir {
-                err,
-                path: dir.clone(),
-            }),
-            _ => Ok(()),
-        };
+        let working_dir = &config.service_working_dir;
+        let root_tmp_dir = &config.service_base_dir.join(&service_id);
 
-        let base_dir = &config.service_base_dir;
-        let service_dir = base_dir.join(&service_id);
-        create(&service_dir)?;
-        create(&service_dir.join(SERVICE_LOCAL_DIR_NAME))?;
-        create(&service_dir.join(SERVICE_TMP_DIR_NAME))?;
+        let service_local_dir = root_tmp_dir.join(SERVICE_LOCAL_DIR_NAME);
+        let service_tmp_dir = root_tmp_dir.join(SERVICE_TMP_DIR_NAME);
 
-        // files will be mapped to service_dir later, along with user-defined ones
+        create(working_dir)?;
+        create(root_tmp_dir)?;
+        create(&service_tmp_dir)?;
+        create(&service_local_dir)?;
+
+        // Special directories that are mapped to service_tmp_dir.
+        // Override user-defined ones.
         let mapped_dirs = hashmap! {
-            format!("/{SERVICE_LOCAL_DIR_NAME}") => PathBuf::from(SERVICE_LOCAL_DIR_NAME),
-            format!("/{SERVICE_TMP_DIR_NAME}") => PathBuf::from(SERVICE_LOCAL_DIR_NAME),
+            format!("{SERVICE_LOCAL_DIR_NAME}") => service_local_dir.clone(),
+            format!("{SERVICE_TMP_DIR_NAME}") => service_tmp_dir.clone(),
+            format!("/{SERVICE_LOCAL_DIR_NAME}") => service_local_dir,
+            format!("/{SERVICE_TMP_DIR_NAME}") => service_tmp_dir,
         };
 
         envs.insert(
@@ -163,14 +164,18 @@ impl<WB: WasmBackend> AppService<WB> {
 
         for module in &mut config.marine_config.modules_config {
             module.config.extend_wasi_envs(envs.clone());
+            // Moves app preopened files and mapped dirs to the &working dir, keeping old aliases.
+            module.config.root_wasi_files_at(working_dir);
+            // Adds /tmp and /local to wasi.
+            // It is important to do it after rooting preopens at working dir, because /tmp and /local are in a separate temporary dir
             module
                 .config
                 .extend_wasi_files(<_>::default(), mapped_dirs.clone());
-            // Must be the last modification of the module.config.
-            // Moves app preopened files and mapped dirs to the &service dir, keeping old aliases.
-            module.config.set_wasi_fs_root(&service_dir)
-        }
 
+            // Create all mapped directories if they do not exist
+            // Needed to provide ability to run the same services both in mrepl and rust-peer
+            create_wasi_dirs(&module.config)?;
+        }
         Ok(())
     }
 
@@ -246,5 +251,30 @@ impl<WB: WasmBackend> AppService<WB> {
         self.marine
             .module_wasi_state(module_name)
             .map_err(Into::into)
+    }
+}
+
+fn create_wasi_dirs(config: &MarineModuleConfig) -> Result<()> {
+    if let Some(wasi_config) = &config.wasi {
+        for dir in wasi_config.mapped_dirs.values() {
+            create(dir)?;
+        }
+
+        for dir in wasi_config.preopened_files.iter() {
+            create(dir)?;
+        }
+    }
+
+    Ok(())
+}
+
+fn create(dir: &Path) -> Result<()> {
+    match std::fs::create_dir_all(dir) {
+        Err(e) if e.kind() == ErrorKind::AlreadyExists => Ok(()),
+        Err(err) => Err(AppServiceError::CreateDir {
+            err,
+            path: dir.to_owned(),
+        }),
+        _ => Ok(()),
     }
 }