+++ marine-js/npm-package/src/__test__/test.spec.ts
@@ -1,11 +1,17 @@
-import fs from 'fs';
-import path from 'path';
-import download from 'download';
-import { MarineService } from '../MarineService';
+import { jest } from '@jest/globals';
+import { defaultImport } from 'default-import';
+import { promises as fsPromises } from 'fs';
+import { createRequire } from 'module';
+import * as path from 'path';
+import * as url from 'url';
+import downloadRaw from 'download';
+import { MarineService } from '../MarineService.js';
 import { callAvm } from '@fluencelabs/avm';
-import { JSONArray, JSONObject } from '../types';
+import { JSONArray, JSONObject } from '../types.js';
 
-const fsPromises = fs.promises;
+const __dirname = url.fileURLToPath(new URL('.', import.meta.url));
+const require = createRequire(import.meta.url);
+const download = defaultImport(downloadRaw);
 
 const vmPeerId = '12D3KooWNzutuy8WHXDKFqFsATvCR6j9cj2FijYbnd47geRKaQZS';