+++ tools/repl/src/repl.rs
@@ -20,10 +20,14 @@ use print_state::print_envs;
 use print_state::print_fs_state;
 use crate::ReplResult;
 
-use fluence_app_service::{DefaultAppService, DefaultWasmBackend, CallParameters, SecurityTetraplet};
+use fluence_app_service::DefaultAppService;
+use fluence_app_service::DefaultWasmBackend;
+use fluence_app_service::CallParameters;
+use fluence_app_service::SecurityTetraplet;
 use fluence_app_service::MarineModuleConfig;
 use fluence_app_service::TomlAppServiceConfig;
 
+use anyhow::anyhow;
 use serde::Deserialize;
 use serde_json::Value as JValue;
 
@@ -64,12 +68,20 @@ struct CallModuleArguments<'args> {
 #[allow(clippy::upper_case_acronyms)]
 pub(super) struct REPL {
     app_service: DefaultAppService,
+    service_working_dir: Option<String>,
 }
 
 impl REPL {
-    pub fn new<S: Into<PathBuf>>(config_file_path: Option<S>, quiet: bool) -> ReplResult<Self> {
-        let app_service = Self::create_app_service(config_file_path, quiet)?;
-        Ok(Self { app_service })
+    pub fn new<S: Into<PathBuf>>(
+        config_file_path: Option<S>,
+        working_dir: Option<String>,
+        quiet: bool,
+    ) -> ReplResult<Self> {
+        let app_service = Self::create_app_service(config_file_path, working_dir.clone(), quiet)?;
+        Ok(Self {
+            app_service,
+            service_working_dir: working_dir,
+        })
     }
 
     /// Returns true, it should be the last executed command.
@@ -96,7 +108,7 @@ impl REPL {
     }
 
     fn new_service<'args>(&mut self, mut args: impl Iterator<Item = &'args str>) {
-        match Self::create_app_service(args.next(), false) {
+        match Self::create_app_service(args.next(), self.service_working_dir.clone(), false) {
             Ok(service) => self.app_service = service,
             Err(e) => println!("failed to create a new application service: {}", e),
         };
@@ -180,7 +192,16 @@ impl REPL {
             {
                 Ok(result) if show_result_arg => {
                     let elapsed_time = start.elapsed();
-                    format!("result: {:?}\n elapsed time: {:?}", result, elapsed_time)
+
+                    let result_string = match serde_json::to_string_pretty(&result) {
+                        Ok(pretty_printed) => pretty_printed,
+                        Err(_) => format!("{:?}", result),
+                    };
+
+                    format!(
+                        "result: {}\n elapsed time: {:?}",
+                        result_string, elapsed_time
+                    )
                 }
                 Ok(_) => {
                     let elapsed_time = start.elapsed();
@@ -222,20 +243,36 @@ impl REPL {
 
     fn create_app_service<S: Into<PathBuf>>(
         config_file_path: Option<S>,
+        working_dir: Option<String>,
         quiet: bool,
     ) -> ReplResult<DefaultAppService> {
         let tmp_path: String = std::env::temp_dir().to_string_lossy().into();
         let service_id = uuid::Uuid::new_v4().to_string();
         let config_file_path: Option<PathBuf> = config_file_path.map(Into::into);
+        let working_dir = working_dir.unwrap_or_else(|| ".".to_string());
 
         let start = Instant::now();
 
         let mut config = config_file_path
             .as_ref()
             .map(TomlAppServiceConfig::load)
-            .transpose()?
+            .transpose()
+            .map_err(|e| {
+                anyhow!(
+                    "failed to load \"{}\": {}",
+                    config_file_path
+                        .as_ref()
+                        .unwrap_or_else(|| panic!(
+                            "config_file_path is Some because it is used to load file"
+                        ))
+                        .display(),
+                    e
+                )
+            })?
             .unwrap_or_default();
+
         config.service_base_dir = Some(tmp_path);
+        config.service_working_dir = Some(working_dir);
 
         config.toml_marine_config.base_path = config_file_path
             .and_then(|path| path.parent().map(PathBuf::from))