+++ crates/fluence-app-service/CHANGELOG.md
@@ -4,6 +4,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+* The following workspace dependencies were updated
+  * dependencies
+    * marine-runtime bumped from 0.24.0 to 0.24.1
+
+## [0.24.0](https://github.com/fluencelabs/marine/compare/fluence-app-service-v0.23.1...fluence-app-service-v0.24.0) (2023-03-03)
+
+
+### ⚠ BREAKING CHANGES
+
+* **fluence-app-service:** add separate bases for temp dirs and mapped dirs ([#288](https://github.com/fluencelabs/marine/issues/288))
+
+### Features
+
+* **fluence-app-service:** add separate bases for temp dirs and mapped dirs ([#288](https://github.com/fluencelabs/marine/issues/288)) ([1d86899](https://github.com/fluencelabs/marine/commit/1d868992bd944eb83926c18a17a24d135c692b4c))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * marine-runtime bumped from 0.24.1 to 0.25.0
+
 ## [Unreleased]
 
 ## [0.23.0] - 2022-12-06