+++ marine/CHANGELOG.md
@@ -4,6 +4,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.25.0](https://github.com/fluencelabs/marine/compare/marine-runtime-v0.24.1...marine-runtime-v0.25.0) (2023-03-03)
+
+
+### ⚠ BREAKING CHANGES
+
+* **fluence-app-service:** add separate bases for temp dirs and mapped dirs ([#288](https://github.com/fluencelabs/marine/issues/288))
+
+### Features
+
+* **fluence-app-service:** add separate bases for temp dirs and mapped dirs ([#288](https://github.com/fluencelabs/marine/issues/288)) ([1d86899](https://github.com/fluencelabs/marine/commit/1d868992bd944eb83926c18a17a24d135c692b4c))
+
+## [0.24.1](https://github.com/fluencelabs/marine/compare/marine-runtime-v0.24.0...marine-runtime-v0.24.1) (2023-02-20)
+
+
+### Bug Fixes
+
+* **mrepl:** replace fn with func in mrepl output to match aqua syntax ([#284](https://github.com/fluencelabs/marine/issues/284)) ([e4c77a8](https://github.com/fluencelabs/marine/commit/e4c77a8cc4c9963ae74e63504dedbcd227bd7cbf))
+
 ## [Unreleased]
 
 ## [0.24.0] - 2022-12-06