+++ tools/cli/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## [0.13.0](https://github.com/fluencelabs/marine/compare/marine-v0.12.7...marine-v0.13.0) (2023-03-03)
+
+
+### ⚠ BREAKING CHANGES
+
+* **cli:** use sdk dependency version from Cargo.lock instead of Cargo.toml ([#286](https://github.com/fluencelabs/marine/issues/286))
+
+### Features
+
+* **cli:** use sdk dependency version from Cargo.lock instead of Cargo.toml ([#286](https://github.com/fluencelabs/marine/issues/286)) ([fc384a4](https://github.com/fluencelabs/marine/commit/fc384a477c2274c9ebff4968871995935b5d6900))
+
+## [0.12.7](https://github.com/fluencelabs/marine/compare/marine-v0.12.6...marine-v0.12.7) (2023-02-10)
+
+
+### Bug Fixes
+
+* **cli:** Make checking latest version an optional feature [#278](https://github.com/fluencelabs/marine/issues/278) ([da31cbb](https://github.com/fluencelabs/marine/commit/da31cbbe38e884ec7989c86af6ebf0fc19093341))
+
 ## [0.12.6](https://github.com/fluencelabs/marine/compare/marine-v0.12.5...marine-v0.12.6) (2023-01-27)