+++ tools/repl/CHANGELOG.md
@@ -4,6 +4,46 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+* The following workspace dependencies were updated
+  * dependencies
+    * fluence-app-service bumped from 0.23.0 to 0.23.1
+
+## [0.20.0](https://github.com/fluencelabs/marine/compare/mrepl-v0.19.2...mrepl-v0.20.0) (2023-03-03)
+
+
+### ⚠ BREAKING CHANGES
+
+* **fluence-app-service:** add separate bases for temp dirs and mapped dirs ([#288](https://github.com/fluencelabs/marine/issues/288))
+
+### Features
+
+* **fluence-app-service:** add separate bases for temp dirs and mapped dirs ([#288](https://github.com/fluencelabs/marine/issues/288)) ([1d86899](https://github.com/fluencelabs/marine/commit/1d868992bd944eb83926c18a17a24d135c692b4c))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * fluence-app-service bumped from 0.23.1 to 0.24.0
+
+## [0.19.1](https://github.com/fluencelabs/marine/compare/mrepl-v0.19.0...mrepl-v0.19.1) (2023-02-10)
+
+
+### Bug Fixes
+
+* **cli:** Make checking latest version an optional feature [#278](https://github.com/fluencelabs/marine/issues/278) ([da31cbb](https://github.com/fluencelabs/marine/commit/da31cbbe38e884ec7989c86af6ebf0fc19093341))
+
+## [0.19.0](https://github.com/fluencelabs/marine/compare/mrepl-v0.18.8...mrepl-v0.19.0) (2023-02-08)
+
+
+### ⚠ BREAKING CHANGES
+
+* add pretty-print for repl output + small error messaging improvement ([#274](https://github.com/fluencelabs/marine/issues/274))
+
+### Features
+
+* add pretty-print for repl output + small error messaging improvement ([#274](https://github.com/fluencelabs/marine/issues/274)) ([9c1f20b](https://github.com/fluencelabs/marine/commit/9c1f20b8a74e467f2e403d18aabb7428baef9bc1))
+
 ## [Unreleased]
 
 ## [0.18.8] - 2022-12-06